Added Fonts, theming and a Welcome View
Fonts chosen are 'Rancho' and 'Inter'. 'Rancho' will be used for the RestorApp logo
and 'Inter' has been set as default for the rest of the app. I have downloaded the assets
and added them to the repo so that Flutter doesn't have to download them on the fly on every run.

I've created my own theme by swapping out the primary and secondary colors as per my
Figma draft, which you can find here:
https://www.figma.com/file/TMaXIRGYD8QId9YgPLj1eT/RestorApp-Library\?node-id\=0%3A1 and
https://www.figma.com/file/VdN5HkgOpmFBTXEfkkOA28/Documentation-App\?node-id\=6%3A5